+++ src/commands/appdistribution-group-create.ts
@@ -6,6 +6,7 @@ import { getProjectName } from "../appdistribution/options-parser-util";
 
 export const command = new Command("appdistribution:group:create <displayName> [alias]")
   .description("create group in project")
+  .alias("appdistribution:groups:create")
   .before(requireAuth)
   .action(async (displayName: string, alias?: string, options?: any) => {
     const projectName = await getProjectName(options);